feat: migrate dashboard & config pages to server components with React Query hydration
