+++ lib/hooks/useUpdateServerConfiguration.ts
@@ -7,7 +7,10 @@ import { apiFetch, useAuthHeader } from "@/lib/api";
 
 type ServerConfiguration = z.infer<typeof ServerConfigurationSchema>;
 
-const updateServerConfiguration = async (config: ServerConfiguration, auth: ReturnType<typeof useAuthHeader>) => {
+const updateServerConfiguration = async (
+  config: ServerConfiguration,
+  auth: ReturnType<typeof useAuthHeader>
+) => {
   return apiFetch(`/server-configuration`, {
     method: "PUT",
     headers: { "Content-Type": "application/json" },
@@ -17,23 +20,43 @@ const updateServerConfiguration = async (config: ServerConfiguration, auth: Retu
 };
 
 export const useUpdateConfiguration = () => {
-  const queryClient = useQueryClient();
+  const qc = useQueryClient();
   const { session } = useSupabaseSession();
   const userId = session?.user.id;
+  const key = ["config", userId ?? "anon"] as const;
 
   const auth = useAuthHeader();
   return useMutation({
-    mutationFn: (conf: ServerConfiguration) => updateServerConfiguration(conf, auth),
-    onSuccess: () => {
+    mutationFn: (next: ServerConfiguration) => updateServerConfiguration(next, auth),
+
+    async onMutate(next) {
+      await qc.cancelQueries({ queryKey: key as unknown as string[] });
+      const prev = qc.getQueryData<ServerConfiguration | null>(key as unknown as string[]);
+      qc.setQueryData(key as unknown as string[], next); // optimistic UI
+      return { prev };
+    },
+
+    onError: (err, _next, ctx) => {
+      if (ctx?.prev !== undefined) {
+        qc.setQueryData(key as unknown as string[], ctx.prev);
+      }
+      toast.error("Save failed", {
+        description: err instanceof Error ? err.message : "An unknown error occurred.",
+      });
+    },
+
+    onSuccess: (saved, next) => {
+      qc.setQueryData(key as unknown as string[], (saved as ServerConfiguration) ?? next);
       toast.success("Configuration saved!", {
         description: "Your changes have been successfully applied.",
       });
-      queryClient.invalidateQueries({ queryKey: ["config", userId] });
     },
-    onError: (err) => {
-      toast.error("Save failed", {
-        description:
-          err instanceof Error ? err.message : "An unknown error occurred.",
+
+    onSettled: () => {
+      qc.invalidateQueries({
+        queryKey: ["config", userId ?? "anon"],
+        exact: true,
+        refetchType: "all",   // refetch active + inactive queries
       });
     },
   });
