+++ lib/types/error.ts
@@ -1,0 +1,8 @@
+export class StatusError extends Error {
+  public readonly status: number;
+
+  constructor(msg: string, status: number) {
+    super(msg);
+    this.status = status;
+  }
+}
