+++ package.json
@@ -28,6 +28,7 @@
     "next-themes": "^0.4.6",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "react-scan": "^0.4.3",
     "sonner": "^2.0.6",
     "sst": "3.17.10",
     "tailwind-merge": "^3.3.1",
